apps/api/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "api" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/apps/api/html' --cov-report xml:'../../coverage/apps/api/coverage.xml' --html='../../reports/apps/api/unittests/html/index.html' --junitxml='../../reports/apps/api/unittests/junit.xml'"

[tool.poetry]
name = "api"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "api"

  [tool.poetry.dependencies]
  python = ">=3.8.1,<4.0.0"

  amqp = "5.1.1"
  asgiref = "3.6.0"
  async-timeout = "4.0.2"
  attrs = "22.2.0"
  billiard = "4.1.0"
  certifi = "2023.5.7"
  charset-normalizer = "3.1.0"
  click = "8.1.3"
  click-didyoumean = "0.3.0"
  click-plugins = "1.1.1"
  click-repl = "0.3.0"
  dj-database-url = "1.2.0"
  Django = "4.1.6"
  django-compat = "1.0.15"
  django-constance = "2.9.1"
  django-cors-headers = "3.13.0"
  django-extensions = "3.2.1"
  django-filter = "22.1"
  django-heroku = "0.3.1"
  django-picklefield = "3.1"
  django4-background-tasks = "1.2.7"
  djangorestframework = "3.14.0"
  djangorestframework-simplejwt = "5.2.2"
  docutils = "0.20.1"
  drf-spectacular = "0.25.1"
  gunicorn = "20.1.0"
  idna = "3.4"
  inflection = "0.5.1"
  jsonschema = "4.17.3"
  kombu = "5.3.1"
  prompt-toolkit = "3.0.39"
  psycopg2 = "2.9.5"
  pydotplus = "2.0.2"
  PyJWT = "2.6.0"
  pyparsing = "3.0.9"
  pyrsistent = "0.19.3"
  python-dateutil = "2.8.2"
  python-dotenv = "1.0.0"
  pytz = "2022.7.1"
  PyYAML = "6.0"
  redis = "4.5.4"
  requests = "2.31.0"
  sentry-sdk = "1.24.0"
  six = "1.16.0"
  sqlparse = "0.4.3"
  typing_extensions = "4.7.1"
  tzdata = "2023.3"
  uritemplate = "4.1.1"
  urllib3 = "1.26.16"
  vine = "5.0.0"
  wcwidth = "0.2.6"
  whitenoise = "6.3.0"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
